applying patches from tracker (5 Issues working with files bigger than 4Gb - ID: 3312690) - patch 03-IntOverflow-grid_resolution.patch

+++ saga-gis/src/modules/grid/grid_gridding/nn/nn.h
@@ -281,7 +281,7 @@ typedef struct nnai nnai;
  * @param d Delaunay triangulation
  * @return Natural Neighbours interpolation
  */
-nnai* nnai_build(delaunay* d, int n, double* x, double* y);
+nnai* nnai_build(delaunay* d, long n, double* x, double* y);
 
 /** Destroys Natural Neighbours array interpolator.
  *
