+++ saga-gis/src/modules/grid/grid_spline/csa.h
@@ -47,7 +47,7 @@ void csa_destroy(csa* a);
 void csa_addpoints(csa* a, int n, point points[]);
 void csa_calculatespline(csa* a);
 void csa_approximate_point(csa* a, point* p);
-void csa_approximate_points(csa* a, int n, point* points);
+void csa_approximate_points(csa* a, long n, point* points);
 
 void csa_setnpmin(csa* a, int npmin);
 void csa_setnpmax(csa* a, int npmax);
